Refactoring server code again
